refactor(backend): conversion to MVC structure
separation of models and controllers in separate folders
