+++ controllers/search_controller.py
@@ -1,10 +1,10 @@
 
 from flask import Flask, request, Response
-from table import Product, Base
+from models.table_model import Product, Base
 import json
 import requests
 import math
-import search_model
+import models.search_model as search_model
 
 class Search_Controller:
     def __init__(self):
