+++ models/category_model.py
@@ -1,6 +1,6 @@
-from database_model import Database_Model
+from models.database_model import Database_Model
 from flask import Flask, request, Response
-from table import Product, Base
+from models.table_model import Product, Base
 import json
 import requests
 import math
